Setup a fairly conservative home-manager
Can't really rely 100% on Nix for all of my software because I'm
running on Linux/BSDs (other than macOS) and it's just annoyances with
dealing with non-Nix software (which is still being worked on).

Couldn't use NixOS because of specific needs and various other
oddities with my main machine.

Right now solely using it for development purposes and on a
per-project basis.